Refactor requireInputs and add new JWT middleware
