+++ back-end/utils/env_config.json
@@ -1,0 +1,4 @@
+{
+  "jwtSecretKey": "fk41*$&@vmqw;s"
+}
+
